+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 - Explicitly add ``gwcs`` to the list of dependencies. [#108]
 - Remove the unused ``stnode_test`` module. [#110]
+- Add support for non-VOUnits to be used by Roman. [#109]
+
 
 0.13.0 (2022-08-23)
 ===================